--- a/packages/start-bun/README.md
+++ b/packages/start-bun/README.md
@@ -1,0 +1,5 @@
+# start-bun
+
+Adapter for Solid apps that builds a Bun server.
+
+This is very experimental; the adapter API isn't at all fleshed out, and things will definitely change.