--- a/packages/start-bun/package.json
+++ b/packages/start-bun/package.json
@@ -7,16 +7,10 @@
     "type": "adapter"
   },
   "dependencies": {
-    "@elysiajs/static": "^0.6.0",
     "@rollup/plugin-commonjs": "^24.1.0",
     "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-node-resolve": "^15.1.0",
-    "compression": "^1.7.4",
-    "elysia": "^0.6.22",
-    "polka": "1.0.0-next.22",
-    "rollup": "^3.26.2",
-    "sirv": "^2.0.3",
-    "terser": "^5.19.0"
+    "rollup": "^3.26.2"
   },
   "devDependencies": {
     "bun-types": "^1.0.1",