--- a/packages/start/node/globals.js
+++ b/packages/start/node/globals.js
@@ -1,7 +1,14 @@
 import crypto from "crypto";
 import Streams from "stream/web";
 
-Object.assign(globalThis, Streams);
+// Bun does not support this assignment.
+// We only assign globalThis if the runtime
+// is not Bun.
+//
+// https://bun.sh/guides/util/detect-bun
+if (!process.versions.bun) {
+  Object.assign(globalThis, Streams);
+}
 
 if (globalThis.crypto != crypto.webcrypto) {
   // @ts-ignore