--- a/packages/start/vite/plugin.js
+++ b/packages/start/vite/plugin.js
@@ -632,6 +632,19 @@ const findAny = (path, name, exts = [".js", ".ts", ".jsx", ".tsx", ".mjs", ".mts
   return null;
 };
 
+const findAdapter = () => {
+  if (process.env.START_ADAPTER) {
+    return process.env.START_ADAPTER;
+  }
+
+  // https://bun.sh/guides/util/detect-bun
+  if (process.versions.bun) {
+    return "solid-start-bun";
+  }
+
+  return "solid-start-node";
+}
+
 /**
  * @param {import('./plugin').Options} options
  * @returns {import('vite').PluginOption[]}
@@ -639,7 +652,7 @@ const findAny = (path, name, exts = [".js", ".ts", ".jsx", ".tsx", ".mjs", ".mts
 export default function solidStart(options) {
   options = Object.assign(
     {
-      adapter: process.env.START_ADAPTER ? process.env.START_ADAPTER : "solid-start-node",
+      adapter: findAdapter(),
       appRoot: "src",
       routesDir: "routes",
       ssr: