Experimental but works 🚀

![image](https://github.com/solidjs/solid-start/assets/28441561/a988d2ac-7b59-4db4-aa5c-f940d1f27021)

## TODOs

- [x] `bun run --bun build`
- [x] `bun run --bun start`
- [x] `bun run --bun dev`
- [ ] Support clientIP (https://github.com/oven-sh/bun/pull/4559)

## Usage

The `--bun` flag is essential; otherwise, this server will be spawned by Node.

```bash
export START_ADAPTER="solid-start-bun"

bun run --bun build
bun run --bun start
```

## Benchmark

wrk command: `wrk -t8 -c500 -d10s --latency "http://localhost:3000"`
system: macOS 13.5.2, on Apple M1.

| Platform | Base RAM usage | Requests/sec | RAM usage after `wrk` |
| --- | --- | --- | --- |
| Node | 22.2MB | 3107.68 | 126.5MB |
| Bun | 16.8MB | 7057.49 | 112.2MB |