@hardywu 2023-10-08
Found a temporary workaround solution, i.e. installing and running the solid-start env with `npm` or `pnpm` ( any env that is not `bun` ).  
Build the `bun` production codes in `./dist` use the bun adapter. 
What we get could be deployed with bun with no issues.